src/requirements.txt
aiohttp
aiohttp_cors
asyncpg
jsonschema==3.0.2
pyyaml
requests
